+++ Services/PriceService.js
@@ -1,12 +1,12 @@
 const ObjectId = require('mongodb').ObjectId
-const OrderService = require('../Services/OrderService')
+const DishesService = require('../Services/DishesService')
 
 const calculateTotalPrice = async (db, finalisedOrder) => {
     let totalPrice = 0
     if(finalisedOrder.orderItems) {
         for (const orderItem of finalisedOrder.orderItems) {
             const dishId = ObjectId(orderItem.menuItemId)
-            const dishPrice = await OrderService.getDishPriceById(db, dishId)
+            const dishPrice = await DishesService.getDishPriceById(db, dishId)
             const totalItemPrice = ((dishPrice * 100 * orderItem.quantity) / 100)
             totalPrice += totalItemPrice
         }
